+++ src/lib.rs
@@ -43,9 +43,9 @@ pub use self::event::{
     BlockListTls, BlockListTlsFields, CryptocurrencyMiningPool, Direction, DnsCovertChannel,
     DomainGenerationAlgorithm, Event, EventDb, EventFilter, EventIterator, EventMessage,
     ExternalDdos, ExtraThreat, FilterEndpoint, FlowKind, FtpBruteForce, FtpPlainText, HttpThreat,
-    LdapBruteForce, LdapPlainText, LearningMethod, MultiHostPortScan, NetworkThreat, NetworkType,
-    NonBrowser, PortScan, RdpBruteForce, RecordType, RepeatedHttpSessions, TorConnection,
-    TrafficDirection, TriageScore, WindowsThreat,
+    LdapBruteForce, LdapPlainText, LearningMethod, LockyRansomware, MultiHostPortScan,
+    NetworkThreat, NetworkType, NonBrowser, PortScan, RdpBruteForce, RecordType,
+    RepeatedHttpSessions, TorConnection, TrafficDirection, TriageScore, WindowsThreat,
 };
 pub use self::migration::{migrate_backend, migrate_data_dir};
 pub use self::model::{Digest as ModelDigest, Model};