+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added new `LockyRansomware` detection event.
+
 ## [0.27.0] - 2024-04-02
 
 ### Added
@@ -904,6 +910,7 @@ leading to a more streamlined system.
 
 - An initial version.
 
+[Unreleased]: https://github.com/petabi/review-database/compare/0.27.0...main
 [0.27.0]: https://github.com/petabi/review-database/compare/0.26.0...0.27.0
 [0.26.0]: https://github.com/petabi/review-database/compare/0.25.0...0.26.0
 [0.25.0]: https://github.com/petabi/review-database/compare/0.24.0...0.25.0